Rename spanpointers.rs to span_pointers.rs
