+++ bottlecap/src/traces/span_pointers.rs
